Codechange: split implementations of ParagraphLayouterFactory into their own file
